+++ lizard_blockbox/management/commands/set_permissions_pdf.py
@@ -1,0 +1,11 @@
+from django.core.management.base import BaseCommand
+
+from lizard_blockbox import import_helpers
+
+
+class Command(BaseCommand):
+    args = ''
+    help = ("Set read permissions for the facsheets.")
+
+    def handle(self, *args, **kwargs):
+        import_helpers.set_permissions_pdf(self.stdout)